feat: add solutions to lcp problem: No.52
No.52.二叉搜索树染色
