+++ lcp/LCP 52. 二叉搜索树染色/Solution.ts
@@ -1,232 +1,3 @@
-# [LCP 52. 二叉搜索树染色](https://leetcode.cn/problems/QO5KpG)
-
-## 题目描述
-
-<!-- 这里写题目描述 -->
-
-欢迎各位勇者来到力扣城，本次试炼主题为「二叉搜索树染色」。
-
-每位勇士面前设有一个**二叉搜索树**的模型，模型的根节点为 `root`，树上的各个节点值均不重复。初始时，所有节点均为蓝色。现在按顺序对这棵二叉树进行若干次操作， `ops[i] = [type, x, y]` 表示第 `i` 次操作为：
-
--   `type` 等于 0 时，将节点值范围在 `[x, y]` 的节点均染蓝
--   `type` 等于 1 时，将节点值范围在 `[x, y]` 的节点均染红
-
-请返回完成所有染色后，该二叉树中红色节点的数量。
-
-**注意：**
-
--   题目保证对于每个操作的 `x`、`y` 值定出现在二叉搜索树节点中
-
-**示例 1：**
-
-> 输入：`root = [1,null,2,null,3,null,4,null,5], ops = [[1,2,4],[1,1,3],[0,3,5]]`
->
-> 输出：`2`
->
-> 解释：
-> 第 0 次操作，将值为 2、3、4 的节点染红；
-> 第 1 次操作，将值为 1、2、3 的节点染红；
-> 第 2 次操作，将值为 3、4、5 的节点染蓝；
-> 因此，最终值为 1、2 的节点为红色节点，返回数量 2
-> ![image.png](https://fastly.jsdelivr.net/gh/doocs/leetcode@main/lcp/LCP%2052.%20%E4%BA%8C%E5%8F%89%E6%90%9C%E7%B4%A2%E6%A0%91%E6%9F%93%E8%89%B2/images/1649833948-arSlXd-image.png)
-
-**示例 2：**
-
-> 输入：`root = [4,2,7,1,null,5,null,null,null,null,6]` > `ops = [[0,2,2],[1,1,5],[0,4,5],[1,5,7]]`
->
-> 输出：`5`
->
-> 解释：
-> 第 0 次操作，将值为 2 的节点染蓝；
-> 第 1 次操作，将值为 1、2、4、5 的节点染红；
-> 第 2 次操作，将值为 4、5 的节点染蓝；
-> 第 3 次操作，将值为 5、6、7 的节点染红；
-> 因此，最终值为 1、2、5、6、7 的节点为红色节点，返回数量 5
-> ![image.png](https://fastly.jsdelivr.net/gh/doocs/leetcode@main/lcp/LCP%2052.%20%E4%BA%8C%E5%8F%89%E6%90%9C%E7%B4%A2%E6%A0%91%E6%9F%93%E8%89%B2/images/1649833763-BljEbP-image.png)
-
-**提示：**
-
--   `1 <= 二叉树节点数量 <= 10^5`
--   `1 <= ops.length <= 10^5`
--   `ops[i].length == 3`
--   `ops[i][0]` 仅为 `0` or `1`
--   `0 <= ops[i][1] <= ops[i][2] <= 10^9`
--   `0 <= 节点值 <= 10^9`
-
-## 解法
-
-<!-- 这里可写通用的实现逻辑 -->
-
-**方法一：有序集合 + 倒序遍历**
-
-我们先遍历一遍二叉树，将所有节点的值加入到有序集合中。
-
-接下来，由于节点的染色状态只跟最后一个操作有关，因此，我们可以倒序遍历操作数组。对于每个操作 $(t, x, y)$，我们在有序集合中找到所有值在 $[x, y]$ 范围内的节点，将其从有序集合中删除。如果 $t = 1$，说明这些节点被染成了红色，我们将其数量累加到答案种。
-
-最后，我们返回答案。
-
-时间复杂度 $O(m + n \times \log n)$，空间复杂度 $O(n)$。其中 $m$ 和 $n$ 分别是操作数组的长度和二叉树的节点数量。
-
-<!-- tabs:start -->
-
-### **Python3**
-
-<!-- 这里可写当前语言的特殊实现逻辑 -->
-
-```python
-# Definition for a binary tree node.
-# class TreeNode:
-#     def __init__(self, x):
-#         self.val = x
-#         self.left = None
-#         self.right = None
-
-from sortedcontainers import SortedList
-
-
-class Solution:
-    def getNumber(self, root: Optional[TreeNode], ops: List[List[int]]) -> int:
-        def dfs(root):
-            if root is None:
-                return
-            sl.add(root.val)
-            dfs(root.left)
-            dfs(root.right)
-
-        sl = SortedList()
-        dfs(root)
-        ans = 0
-        for t, x, y in ops[::-1]:
-            i = sl.bisect_left(x)
-            while i < len(sl) and sl[i] <= y:
-                sl.pop(i)
-                ans += t == 1
-        return ans
-```
-
-### **Java**
-
-<!-- 这里可写当前语言的特殊实现逻辑 -->
-
-```java
-/**
- * Definition for a binary tree node.
- * public class TreeNode {
- *     int val;
- *     TreeNode left;
- *     TreeNode right;
- *     TreeNode(int x) { val = x; }
- * }
- */
-class Solution {
-    private TreeSet<Integer> ts = new TreeSet<>();
-
-    public int getNumber(TreeNode root, int[][] ops) {
-        dfs(root);
-        int ans = 0;
-        for (int i = ops.length - 1; i >= 0; --i) {
-            int t = ops[i][0];
-            int x = ops[i][1], y = ops[i][2];
-            Integer ceiling = ts.ceiling(x);
-            while (ceiling != null && ceiling <= y) {
-                ts.remove(ceiling);
-                ceiling = ts.ceiling(x);
-                ans += t;
-            }
-        }
-        return ans;
-    }
-
-    private void dfs(TreeNode root) {
-        if (root == null) {
-            return;
-        }
-        ts.add(root.val);
-        dfs(root.left);
-        dfs(root.right);
-    }
-}
-```
-
-### **C++**
-
-```cpp
-/**
- * Definition for a binary tree node.
- * struct TreeNode {
- *     int val;
- *     TreeNode *left;
- *     TreeNode *right;
- *     TreeNode(int x) : val(x), left(NULL), right(NULL) {}
- * };
- */
-class Solution {
-public:
-    int getNumber(TreeNode* root, vector<vector<int>>& ops) {
-        set<int> ts;
-        function<void(TreeNode*)> dfs = [&](TreeNode* root) {
-            if (root == nullptr) {
-                return;
-            }
-            ts.insert(root->val);
-            dfs(root->left);
-            dfs(root->right);
-        };
-        dfs(root);
-        int ans = 0;
-        for (int i = ops.size() - 1; ~i; --i) {
-            int t = ops[i][0];
-            int x = ops[i][1], y = ops[i][2];
-            auto it = ts.lower_bound(x);
-            while (it != ts.end() && *it <= y) {
-                ts.erase(it++);
-                ans += t;
-            }
-        }
-        return ans;
-    }
-};
-```
-
-### **Go**
-
-```go
-/**
- * Definition for a binary tree node.
- * type TreeNode struct {
- *     Val int
- *     Left *TreeNode
- *     Right *TreeNode
- * }
- */
-func getNumber(root *TreeNode, ops [][]int) (ans int) {
-	rbt := redblacktree.NewWithIntComparator()
-	var dfs func(root *TreeNode)
-	dfs = func(root *TreeNode) {
-		if root == nil {
-			return
-		}
-		rbt.Put(root.Val, nil)
-		dfs(root.Left)
-		dfs(root.Right)
-	}
-	dfs(root)
-	for i := len(ops) - 1; i >= 0; i-- {
-		t, x, y := ops[i][0], ops[i][1], ops[i][2]
-		node, _ := rbt.Ceiling(x)
-		for node != nil && node.Key.(int) <= y {
-			rbt.Remove(node.Key)
-			node, _ = rbt.Ceiling(x)
-			ans += t
-		}
-	}
-	return
-}
-```
-
-### **TypeScript**
-
-```ts
 /**
  * Definition for a binary tree node.
  * class TreeNode {
@@ -758,12 +529,3 @@ class TreeSet<T = number> {
         return undefined;
     }
 }
-```
-
-### **...**
-
-```
-
-```
-
-<!-- tabs:end -->
